funkcne fotky a pop-up uz aj v search listing plus opravene pridavanie fotiek aby nepridavalo do databazu aj ked uzivatel fotku nepridal

+++ www/addNew.php
@@ -130,14 +130,15 @@ if(isset($_POST["addNewAd"])) {
         $var3 = $var1.$var2;  // concatenate $var1 and $var2 in $var3
         $var3 = md5($var3);   // convert $var3 using md5 function and generate 32 characters hex number
 
-        $fnm = $_FILES["image"]["name"];    // get the image name in $fnm variable
-        $dst = "./images/".$var3.$fnm;  // storing image path into the {all_images} folder with 32 characters hex number and file name
-        $dst_db = "images/".$var3.$fnm; // storing image path into the database with 32 characters hex number and file name
+        if(file_exists($_FILES["image"]["tmp_name"]) || is_uploaded_file($_FILES["image"]["tmp_name"])) { //vykona sa iba ak uzivatel zadal subor
+            $fnm = $_FILES["image"]["name"];    // get the image name in $fnm variable
+            $dst = "./images/".$var3.$fnm;  // storing image path into the {all_images} folder with 32 characters hex number and file name
+            $dst_db = "images/".$var3.$fnm; // storing image path into the database with 32 characters hex number and file name
 
-        move_uploaded_file($_FILES["image"]["tmp_name"],$dst);  // move image into the {all_images} folder with 32 characters hex number and image name
+            move_uploaded_file($_FILES["image"]["tmp_name"],$dst);  // move image into the {all_images} folder with 32 characters hex number and image name
 
-        //$check = mysqli_query($db,"insert into tbltest(fname,images) values('$_POST[fname]','$dst_db')");  // executing insert query
-        $check = $storage->insertImage($dst_db);  // executing insert query
+            $check = $storage->insertImage($dst_db);  // executing insert query
+        }
         //koniec kopcenia
         ?>
         <script>
